Use a fork of okhttp
This fork of RN was made for a project that was crashing on special
characters in cookies. The cookies are received in the response of a
backend call. We don't maintain the backend, so we couldn't fix it there.
Every cookie that is received, is saved in a 'CookieJar', these cookies
are being copied to all folowing requests to that same domain. Before
doing a new request, okhttp checks all headers and only a certain
character range is accepted. If any header contains a char outside of
that character range, an exception is thrown. As we don't do the calls
ourselves, we can't add a try/catch.
This problem only occurs on Android.